draft of serenity repository
